refactor: relocate test-user setup out of lib into the e2e suite
